Music album methods-restructure
