+++ main.rb
@@ -1,7 +1,7 @@
-# rubocop:disable Layout/LineLength
 require_relative 'preserve_data'
 require 'json'
 require 'date'
+require_relative 'music_helper'
 require_relative 'item'
 require_relative 'author'
 require_relative 'book'
@@ -10,7 +10,6 @@ require_relative 'genre'
 require_relative 'label'
 require_relative 'music_album'
 require_relative 'app'
-# rubocop:enable Layout/LineLength
 app = App.new
 
 # Store your items
@@ -95,66 +94,7 @@ end
 
 # adding game   [END]........................................................................................
 
-# save music_album   [START]>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
-def add_music_album(music_albums, items, genres)
-  puts 'Adding a new music album...'
-  print 'Enter album name: '
-  album_name = gets.chomp.to_s
-  print 'Enter genre name: '
-  genre_name = gets.chomp.to_s
-  print 'Is the album on Spotify? (true or false): '
-  on_spotify = gets.chomp.downcase == 'true'
-  print 'Can the album be archived? (true or false): '
-  can_be_archived = gets.chomp.downcase == 'true'
-
-  genres = load_genres
-
-  genre = genres.find { |g| g.name == genre_name }
-  unless genre
-    genre = Genre.new(genres.size + 1, genre_name)
-    genres << genre
-    save_genres(genres)
-  end
-
-  new_id = music_albums.size + 1
-  music_album = MusicAlbum.new(new_id, album_name, on_spotify, can_be_archived, genre)
-
-  music_album.can_be_archived = can_be_archived
-  items << music_album
-  music_albums << { 'id' => new_id, 'album' => { 'can_be_archived' => can_be_archived, 'on_spotify' => on_spotify, 'album_name' => album_name, 'genre_name' => genre_name } }
-
-  save_music_albums(music_albums)
-  puts 'Music album added successfully.'
-end
-
-# Save music albums to the JSON file
-
-def save_music_albums(music_albums)
-  data_to_save = music_albums.map do |album_data|
-    {
-      'id' => album_data['id'],
-      'album' => {
-        'can_be_archived' => album_data['album']['can_be_archived'],
-        'on_spotify' => album_data['album']['on_spotify'],
-        'album_name' => album_data['album']['album_name'],
-        # 'genre_id' => album_data['album']['genre_id']
-        'genre_name' => album_data['album']['genre_name']
-      }
-    }
-  end
-
-  json_file_path = 'music.json'
-
-    File.open(json_file_path, 'w') do |file|
-    file.write(JSON.pretty_generate(data_to_save))
-  end
-
-  puts 'Music albums saved successfully.'
-rescue JSON::GeneratorError => e
-  puts "Error generating JSON data for '#{json_file_path}': #{e.message}"
-rescue StandardError => e
-  puts "Error saving music albums: #{e.message}"
-end
+# music_album   [START]>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
 
 def save_genres(genres)
   genre_data = genres.map do |genre|
@@ -169,65 +109,33 @@ rescue JSON::GeneratorError => e
   puts "Error generating JSON data for 'genre.json': #{e.message}"
 end
 
-# list methods music_album........
-# list genres from genre.json file
-
-def list_genres(genres)
-  puts 'Listing all genres:'
-  genres.each do |genre|
-    puts genre['genre_name']
-  end
-end
-
-# list music albums
-def list_all_music_albums(music_albums)
-  puts 'Listing all music albums:'
-  music_albums.each_with_index do |album_data, index|
-    display_music_album(index + 1, album_data['album'])
-  end
-end
-
-def display_music_album(index, album)
-  puts "#{index}. Music Album -"
-  display_music_album_details(album)
-end
-
-def display_music_album_details(album)
-  puts "Album Name: #{album['album_name'] || 'No Album Name'}"
-  puts "Can Be Archived: #{album['can_be_archived'] ? 'Yes' : 'No'}"
-  puts "On Spotify: #{album['on_spotify'] ? 'Yes' : 'No'}"
-  genre_name = album['genre'] ? album['genre']['name'] : nil
-  puts "Genre: #{genre_name || 'No Genre'}"
-end
-
 # load and save data to json file [START]......
 def load_music_albums(genres)
   if File.exist?('music.json')
     json_data = File.read('music.json')
 
-  JSON.parse(json_data).map do |album_data|
-  album = album_data['album']
-  id = album_data['id']
-  album_name = album['album_name']
-  can_be_archived = album['can_be_archived']
-  on_spotify = album['on_spotify']
-  genre_data = album['genre']
-  genre_name = genre_data['name'] if genre_data.is_a?(Hash) && genre_data.key?('name')
-  genre = genres.find { |g| g.name == genre_name }
-  unless genre
-    genre = Genre.new(genres.size + 1, genre_name)
-    genres << genre
-  end
-  puts "Loaded music album: ID=#{id}, Name=#{album_name}, Can Be Archived=#{can_be_archived}, On Spotify=#{on_spotify}, Genre=#{genre_name}"
-  MusicAlbum.new(id, album_name, on_spotify, can_be_archived, genre)
-  end
+    JSON.parse(json_data).map do |album_data|
+      album = album_data['album']
+      id = album_data['id']
+      album_name = album['album_name']
+      can_be_archived = album['can_be_archived']
+      on_spotify = album['on_spotify']
+      genre_data = album['genre']
+      genre_name = genre_data['name'] if genre_data.is_a?(Hash) && genre_data.key?('name')
+      genre = genres.find { |g| g.name == genre_name }
+      unless genre
+        genre = Genre.new(genres.size + 1, genre_name)
+        genres << genre
+      end
+      puts "Loaded music album: ID=#{id}, Name=#{album_name}, Can Be Archived=#{can_be_archived}, On Spotify=#{on_spotify}, Genre=#{genre_name}"
+      MusicAlbum.new(id, album_name, on_spotify, can_be_archived, genre)
+    end
     []
   end
 rescue JSON::ParserError => e
   puts "Error parsing 'music.json': #{e.message}"
   []
 end
-
 
 def load_genres
   if File.exist?('genre.json')
